+++ src/v2/index.ts
@@ -1,16 +1,15 @@
-import { Context, getIndexesFromSegment } from "./utils";
+import { Context } from "./utils";
 import { Side } from "../shared/language";
 import { Iterator } from "./iterator";
-import { getBestMatch, getSubSequenceNodes, isLatterCandidateBetter, oneSidedIteration } from "./core";
-import { Change } from "./diff";
-import { DiffType } from "../types";
+import { computeDiff } from "./core";
 import { asciiRenderFn, fail, prettyRenderFn } from "../debug";
-import { range } from "../utils";
 import { Options, OutputType, ResultTypeMapper } from "./types";
 import { applyChangesToSources } from "./printer";
+import { Change } from "./change";
 
-const defaultOptions: Options = {
+const defaultOptions: Required<Options> = {
   outputType: OutputType.changes,
+  tryAlignMoves: true,
 };
 
 export function getDiff2<_OutputType extends OutputType = OutputType.changes>(sourceA: string, sourceB: string, options?: Options<_OutputType>): ResultTypeMapper[_OutputType] {
@@ -19,81 +18,11 @@ export function getDiff2<_OutputType extends OutputType = OutputType.changes>(so
   const iterA = new Iterator(sourceA, Side.a);
   const iterB = new Iterator(sourceB, Side.b);
 
-  // console.log('------------- A ----------')
-  // iterA.nodes.map((x, i) => {
-  //   console.log(i, "|", x.start, x.end, x.prettyKind, `"${x.text}"`)
-  // })
+  let changes: Change[] = [];
 
-  // console.log('------------- B ----------')
-  // iterB.nodes.map((x, i) => {
-  //   console.log(i, "|", x.start, x.end, x.prettyKind, `"${x.text}"`)
-  // })
+  _context = new Context(sourceA, sourceB, iterA, iterB, changes);
 
-  const changes: Change[] = [];
-
-  _context = new Context(sourceA, sourceB, changes);
-
-  _context.iterA = iterA;
-  _context.iterB = iterB;
-
-  while (true) {
-    const a = iterA.next();
-    const b = iterB.next();
-
-    // No more nodes to process. We are done
-    if (!a && !b) {
-      break;
-    }
-
-    // One of the iterators ran out of nodes. We will mark the remaining unmatched nodes
-    if (!a || !b) {
-      // If A finished means that B still have nodes, they are additions. If B finished means that A have nodes, they are deletions
-      const iterOn = !a ? iterB : iterA;
-      const type = !a ? DiffType.addition : DiffType.deletion;
-      const startFrom = !a ? b?.index : a.index;
-
-      const remainingChanges = oneSidedIteration(iterOn, type, startFrom!);
-      changes.push(...remainingChanges);
-      break;
-    }
-
-    // 1-
-    const bestMatchForB = getBestMatch(b);
-
-    if (!bestMatchForB) {
-      const addition = Change.createAddition(b);
-      changes.push(addition);
-      iterB.mark(b.index, DiffType.addition);
-
-      continue;
-    }
-
-    // May be empty if the node we are looking for was the only one
-    const subSequenceNodesToCheck = getSubSequenceNodes(bestMatchForB, b);
-
-    let bestCandidate = bestMatchForB;
-
-    for (const node of subSequenceNodesToCheck) {
-      const newCandidate = getBestMatch(node);
-
-      if (!newCandidate) {
-        const addition = Change.createAddition(b);
-        changes.push(addition);
-        iterB.mark(b.index, DiffType.addition);
-
-        continue;
-      }
-
-      if (isLatterCandidateBetter(bestCandidate, newCandidate)) {
-        bestCandidate = newCandidate;
-      }
-    }
-
-    const move = Change.createMove(bestCandidate);
-    changes.push(move);
-    markMatched(move);
-    continue;
-  }
+  changes = computeDiff();
 
   switch (_options.outputType) {
     case OutputType.changes: {
@@ -107,20 +36,6 @@ export function getDiff2<_OutputType extends OutputType = OutputType.changes>(so
     }
     default:
       fail();
-  }
-}
-
-function markMatched(change: Change) {
-  for (const segment of change.segments) {
-    const { a, b } = getIndexesFromSegment(segment);
-
-    for (const index of range(a.start, a.end)) {
-      _context.iterA.mark(index, change.type);
-    }
-
-    for (const index of range(b.start, b.end)) {
-      _context.iterB.mark(index, change.type);
-    }
   }
 }
