+++ src/v2/types.ts
@@ -1,5 +1,4 @@
-import { DiffType } from "../types";
-import { Change } from "./diff";
+import { Change } from "./change";
 import { Node } from "./node";
 
 export type SyntaxKind = number;
@@ -13,7 +12,8 @@ export interface ParsedProgram {
 }
 
 export interface Options<_OutputType extends OutputType = OutputType.changes> {
-  outputType: _OutputType;
+  outputType?: _OutputType;
+  tryAlignMoves?: boolean;
 }
 
 export enum OutputType {
@@ -26,4 +26,13 @@ export interface ResultTypeMapper {
   [OutputType.changes]: Change[];
   [OutputType.text]: { sourceA: string; sourceB: string };
   [OutputType.prettyText]: { sourceA: string; sourceB: string };
+}
+
+// Start is inclusive, end is not inclusive
+export type Segment = [indexA: number, indexB: number, length: number];
+
+export interface CandidateMatch {
+  length: number;
+  skips: number;
+  segments: Segment[];
 }
